Feature gate COUNTERID features.
For two reasons: one, it's not really useful without the ability to
perform lockless reads.  Inexact summary statistics aren't that
interesting to me, though reasonable people can disagree.  Two, the
feature will require significant revisions if if the lockless reads are
to be implemented.  Three, if lockless reads are added, this feature
gate will provide a useful way to check their performance impact.