Support Clone as well as Copy for find.
I believe there's no correctness implication, and it optimizes to the
same code if you use the default Clone implementation.  The primary
reason I'm adding this is to support using a smart pointer like Arc
together with the rest of the hash map by using a non-default Clone
implementation over a raw pointer (thus allowing the
required-for-memory-safety bump of the reference count to occur while
the lock is taken, ensuring that the pointer remains valid).

Note that this isn't a robust solution by any means: in particular,
dropping the map won't drop anything that is being pointed to, since as
far as it's concerned they're just raw pointers.  This is really only
suited to some sort of long-lived (e.g. duration of program) hash map
where the only time you want the contents of a smart pointer dropped is
when it is explicitly deleted from the table.